+++ src/main/java/org/betonquest/betonquest/compatibility/jobsreborn/ObjectiveLeaveJob.java
@@ -23,7 +23,6 @@ public class ObjectiveLeaveJob extends Objective implements Listener {
     @SuppressWarnings("PMD.AvoidLiteralsInIfCondition")
     public ObjectiveLeaveJob(final Instruction instructions) throws InstructionParseException {
         super(instructions);
-        template = ObjectiveData.class;
         if (instructions.size() < 2) {
             throw new InstructionParseException("Not enough arguments");
         }