+++ CHANGELOG.md
@@ -11,6 +11,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - command `variable` to list and modify variables on a variable objective
 - config option `conversation_IO_config.slowtellraw.message_delay` to set the delay between messages in the SlowTellRaw conversation IO
 - `resourcepack` objective - to allow checking when the player accepts, denies, downloads, or fails to download a resource pack 
+- `region` objective now supports variables as region name
+- `location` objective now allows `entry` and `exit` keywords to track entering and leaving the location
 ### Changed
 - player variable
   - `%player%` is now the same as `%player.name%`